Remove structs.go and seperate intent, response and entity
structs.go is very generic and not very meaningful. It’s better to keep response, entities and intents seperated and so they can be tested individually. Their respective structs should be associted with their specific files
