+++ recast/entity.go
@@ -1,5 +1,12 @@
 package recast
 
+// Entity defines the details for a single entity
+type Entity struct {
+	data       map[string]interface{} // Json data
+	Name       string
+	Confidence float64
+}
+
 func NewEntity(name string, jsonData map[string]interface{}) Entity {
 	var e Entity
