+++ recast/intent.go
@@ -1,0 +1,7 @@
+package recast
+
+// Intent defines the details which define a single intent
+type Intent struct {
+	Name       string
+	Confidence float64
+}
